Move icon/chemistry/ to chemistry/icon/
Convert relative imports to absolute, update icon_physics.py.
