+++ jcm/physics/chemistry/icon/__init__.py
